requirements.txt
loguru==0.7.2
better-automation==2.0.2
inquirer==3.2.1
termcolor==2.4.0
fake-useragent==1.4.0
sqlalchemy==2.0.25
web3==6.14.0
aiosqlite==0.19.0
aiofiles==23.2.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0